LangChain/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-google-firestore
langchain_ollama
firestore
chromadb
tiktoken
sentence-transformers
bs4
firecrawl-py
langchainhub
wikipedia
tavily-python
python-dotenv